refactor gov keys
