+++ x/gov/proposal.go
@@ -1,0 +1,172 @@
+package gov
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
+
+// SubmitProposal create new proposal given a content
+func (keeper Keeper) SubmitProposal(ctx sdk.Context, content Content) (Proposal, sdk.Error) {
+	if !keeper.router.HasRoute(content.ProposalRoute()) {
+		return Proposal{}, ErrNoProposalHandlerExists(keeper.codespace, content)
+	}
+
+	// Execute the proposal content in a cache-wrapped context to validate the
+	// actual parameter changes before the proposal proceeds through the
+	// governance process. State is not persisted.
+	cacheCtx, _ := ctx.CacheContext()
+	handler := keeper.router.GetRoute(content.ProposalRoute())
+	if err := handler(cacheCtx, content); err != nil {
+		return Proposal{}, ErrInvalidProposalContent(keeper.codespace, err.Result().Log)
+	}
+
+	proposalID, err := keeper.getNewProposalID(ctx)
+	if err != nil {
+		return Proposal{}, err
+	}
+
+	submitTime := ctx.BlockHeader().Time
+	depositPeriod := keeper.GetDepositParams(ctx).MaxDepositPeriod
+
+	proposal := NewProposal(content, proposalID, submitTime, submitTime.Add(depositPeriod))
+
+	keeper.SetProposal(ctx, proposal)
+	keeper.InsertInactiveProposalQueue(ctx, proposal.DepositEndTime, proposalID)
+
+	return proposal, nil
+}
+
+// GetProposal get Proposal from store by ProposalID
+func (keeper Keeper) GetProposal(ctx sdk.Context, proposalID uint64) (proposal Proposal, ok bool) {
+	store := ctx.KVStore(keeper.storeKey)
+	bz := store.Get(KeyProposal(proposalID))
+	if bz == nil {
+		return
+	}
+	keeper.cdc.MustUnmarshalBinaryLengthPrefixed(bz, &proposal)
+	return proposal, true
+}
+
+// SetProposal set a proposal to store
+func (keeper Keeper) SetProposal(ctx sdk.Context, proposal Proposal) {
+	store := ctx.KVStore(keeper.storeKey)
+	bz := keeper.cdc.MustMarshalBinaryLengthPrefixed(proposal)
+	store.Set(KeyProposal(proposal.ProposalID), bz)
+}
+
+// DeleteProposal deletes a proposal from store
+func (keeper Keeper) DeleteProposal(ctx sdk.Context, proposalID uint64) {
+	store := ctx.KVStore(keeper.storeKey)
+	proposal, ok := keeper.GetProposal(ctx, proposalID)
+	if !ok {
+		panic("DeleteProposal cannot fail to GetProposal")
+	}
+	keeper.RemoveFromInactiveProposalQueue(ctx, proposal.DepositEndTime, proposalID)
+	keeper.RemoveFromActiveProposalQueue(ctx, proposal.VotingEndTime, proposalID)
+	store.Delete(KeyProposal(proposalID))
+}
+
+// GetProposalsFiltered get Proposals from store by ProposalID
+// voterAddr will filter proposals by whether or not that address has voted on them
+// depositorAddr will filter proposals by whether or not that address has deposited to them
+// status will filter proposals by status
+// numLatest will fetch a specified number of the most recent proposals, or 0 for all proposals
+func (keeper Keeper) GetProposalsFiltered(ctx sdk.Context, voterAddr sdk.AccAddress, depositorAddr sdk.AccAddress, status ProposalStatus, numLatest uint64) []Proposal {
+
+	maxProposalID, err := keeper.peekCurrentProposalID(ctx)
+	if err != nil {
+		return nil
+	}
+
+	matchingProposals := []Proposal{}
+
+	if numLatest == 0 {
+		numLatest = maxProposalID
+	}
+
+	for proposalID := maxProposalID - numLatest; proposalID < maxProposalID; proposalID++ {
+		if voterAddr != nil && len(voterAddr) != 0 {
+			_, found := keeper.GetVote(ctx, proposalID, voterAddr)
+			if !found {
+				continue
+			}
+		}
+
+		if depositorAddr != nil && len(depositorAddr) != 0 {
+			_, found := keeper.GetDeposit(ctx, proposalID, depositorAddr)
+			if !found {
+				continue
+			}
+		}
+
+		proposal, ok := keeper.GetProposal(ctx, proposalID)
+		if !ok {
+			continue
+		}
+
+		if ValidProposalStatus(status) {
+			if proposal.Status != status {
+				continue
+			}
+		}
+
+		matchingProposals = append(matchingProposals, proposal)
+	}
+	return matchingProposals
+}
+
+// GetLastProposalID gets the last used proposal ID
+func (keeper Keeper) GetLastProposalID(ctx sdk.Context) (proposalID uint64) {
+	proposalID, err := keeper.peekCurrentProposalID(ctx)
+	if err != nil {
+		return 0
+	}
+	proposalID--
+	return
+}
+
+// Gets the next available ProposalID and increments it
+func (keeper Keeper) getNewProposalID(ctx sdk.Context) (proposalID uint64, err sdk.Error) {
+	store := ctx.KVStore(keeper.storeKey)
+	bz := store.Get(KeyNextProposalID)
+	if bz == nil {
+		return 0, ErrInvalidGenesis(keeper.codespace, "initial proposal ID never set")
+	}
+	keeper.cdc.MustUnmarshalBinaryLengthPrefixed(bz, &proposalID)
+	bz = keeper.cdc.MustMarshalBinaryLengthPrefixed(proposalID + 1)
+	store.Set(KeyNextProposalID, bz)
+	return proposalID, nil
+}
+
+// Set the initial proposal ID
+func (keeper Keeper) setInitialProposalID(ctx sdk.Context, proposalID uint64) sdk.Error {
+	store := ctx.KVStore(keeper.storeKey)
+	bz := store.Get(KeyNextProposalID)
+	if bz != nil {
+		return ErrInvalidGenesis(keeper.codespace, "initial proposal ID already set")
+	}
+	bz = keeper.cdc.MustMarshalBinaryLengthPrefixed(proposalID)
+	store.Set(KeyNextProposalID, bz)
+	return nil
+}
+
+// Peeks the next available ProposalID without incrementing it
+func (keeper Keeper) peekCurrentProposalID(ctx sdk.Context) (proposalID uint64, err sdk.Error) {
+	store := ctx.KVStore(keeper.storeKey)
+	bz := store.Get(KeyNextProposalID)
+	if bz == nil {
+		return 0, ErrInvalidGenesis(keeper.codespace, "initial proposal ID never set")
+	}
+	keeper.cdc.MustUnmarshalBinaryLengthPrefixed(bz, &proposalID)
+	return proposalID, nil
+}
+
+func (keeper Keeper) activateVotingPeriod(ctx sdk.Context, proposal Proposal) {
+	proposal.VotingStartTime = ctx.BlockHeader().Time
+	votingPeriod := keeper.GetVotingParams(ctx).VotingPeriod
+	proposal.VotingEndTime = proposal.VotingStartTime.Add(votingPeriod)
+	proposal.Status = StatusVotingPeriod
+	keeper.SetProposal(ctx, proposal)
+
+	keeper.RemoveFromInactiveProposalQueue(ctx, proposal.DepositEndTime, proposal.ProposalID)
+	keeper.InsertActiveProposalQueue(ctx, proposal.VotingEndTime, proposal.ProposalID)
+}
