+++ x/gov/vote.go
@@ -1,0 +1,58 @@
+package gov
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
+
+// Adds a vote on a specific proposal
+func (keeper Keeper) AddVote(ctx sdk.Context, proposalID uint64, voterAddr sdk.AccAddress, option VoteOption) sdk.Error {
+	proposal, ok := keeper.GetProposal(ctx, proposalID)
+	if !ok {
+		return ErrUnknownProposal(keeper.codespace, proposalID)
+	}
+	if proposal.Status != StatusVotingPeriod {
+		return ErrInactiveProposal(keeper.codespace, proposalID)
+	}
+
+	if !ValidVoteOption(option) {
+		return ErrInvalidVote(keeper.codespace, option)
+	}
+
+	vote := Vote{
+		ProposalID: proposalID,
+		Voter:      voterAddr,
+		Option:     option,
+	}
+	keeper.setVote(ctx, proposalID, voterAddr, vote)
+
+	return nil
+}
+
+// Gets the vote of a specific voter on a specific proposal
+func (keeper Keeper) GetVote(ctx sdk.Context, proposalID uint64, voterAddr sdk.AccAddress) (Vote, bool) {
+	store := ctx.KVStore(keeper.storeKey)
+	bz := store.Get(KeyVote(proposalID, voterAddr))
+	if bz == nil {
+		return Vote{}, false
+	}
+	var vote Vote
+	keeper.cdc.MustUnmarshalBinaryLengthPrefixed(bz, &vote)
+	return vote, true
+}
+
+func (keeper Keeper) setVote(ctx sdk.Context, proposalID uint64, voterAddr sdk.AccAddress, vote Vote) {
+	store := ctx.KVStore(keeper.storeKey)
+	bz := keeper.cdc.MustMarshalBinaryLengthPrefixed(vote)
+	store.Set(KeyVote(proposalID, voterAddr), bz)
+}
+
+// Gets all the votes on a specific proposal
+func (keeper Keeper) GetVotes(ctx sdk.Context, proposalID uint64) sdk.Iterator {
+	store := ctx.KVStore(keeper.storeKey)
+	return sdk.KVStorePrefixIterator(store, KeyVotesSubspace(proposalID))
+}
+
+func (keeper Keeper) deleteVote(ctx sdk.Context, proposalID uint64, voterAddr sdk.AccAddress) {
+	store := ctx.KVStore(keeper.storeKey)
+	store.Delete(KeyVote(proposalID, voterAddr))
+}
